finish

+++ wordpress/start.sh
@@ -1,24 +1,31 @@
 #!/bin/sh
-cd /var/www/html
-
 set -e
 
-wp core download --allow-root
+cd /var/www/html
 
-wp   config create --dbname="${SQL_DATABASE}" \
-                --dbuser="${SQL_USER}" \
-                --dbpass="$(cat ../secrets/database_pass.txt)" \
-                --dbhost="mariadb"
+until mysqladmin ping -h mariadb --silent; do
+    sleep 1
+done
+if [ ! -f wp-config.php ]; then
+    if [ ! -f wp-load.php ]; then
+        wp core download --allow-root
+    fi
+    wp   config create --dbname="${SQL_DATABASE}" \
+                   --dbuser="${SQL_USER}" \
+                    --dbpass="$(cat /run/secrets/database_pass)" \
+                    --dbhost="mariadb"\
+                    --allow-root    
 
+    wp core install --url="https://mfahmi.42.fr" \
+                    --title="${WP_TITLE}" \
+                    --admin_user="${WP_ROOT_USER}" \
+                    --admin_password="$(cat /run/secrets/wp_pass_root)" \
+                    --admin_email="${WP_ADMIN_EMAIL}"\
+                    --allow-root
 
-wp core install --url="https://mfahmi.42.fr" \
-                --title="${WP_TITLE}" \
-                --admin_user="${WP_ADMIN_USER}" \
-                --admin_password="$(cat ../secrets/database_pass_root.txt)" \
-                --admin_email="${WP_ADMIN_EMAIL}"
-
-wp user create "${WP_USER}" "${WP_EMAIL}" \
-    --role="${WP_ROLE}" \
-    --user_pass="$(cat ../secrets/wp_pass.txt)"
-
-exec php-fpm -F
+    wp user create "${WP_USER}" "${WP_EMAIL}" \
+        --role="${WP_ROLE}" \
+        --user_pass="$(cat  /run/secrets/wp_pass)"\
+        --allow-root
+fi
+exec php-fpm8.2 -F
